refactor: initialize 메서드 제거
