+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapterRegistry.java
@@ -1,0 +1,20 @@
+package com.interface21.webmvc.servlet.mvc.tobe;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class HandlerAdapterRegistry {
+    private final List<HandlerAdapter> handlerAdapters = new ArrayList<>();
+
+    public void addHandlerAdapter(HandlerAdapter handlerAdapter) {
+        handlerAdapters.add(handlerAdapter);
+    }
+
+    public HandlerAdapter getHandlerAdapter(Object handler) {
+        return handlerAdapters.stream()
+                .filter(handlerAdapter -> handlerAdapter.supports(handler))
+                .findFirst()
+                .orElseThrow(() -> new RuntimeException("가능한 HandlerAdapter가 존재하지 않습니다."));
+    }
+
+}
